# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auto Tournament API",
    "version": "1.0.0",
    "description": "Run a tournament from a bot or a script: teams, players, brackets, matches and results. The dashboard is a client of this API and nothing more. Which game a tournament runs is a module's business, so the endpoints here are the same whether it is Counter-Strike 2 or a game whose captains report their own scores.",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3069",
      "description": "Your MAT instance. Replace with its base URL."
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Service token for machine clients (bots, scripts, CI). Send `Authorization: Bearer <token>`, where the token is one configured in API_TOKENS (full admin) or API_TOKENS_READONLY (GET only). Human admins authenticate with a session cookie instead."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Service token for machine clients (bots, scripts, CI). Send `Authorization: Bearer <token>`, where the token is one configured in API_TOKENS (full admin) or API_TOKENS_READONLY (GET only). Human admins authenticate with a session cookie instead."
      },
      "apiToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Token",
        "description": "Service token, as an alternative to `Authorization: Bearer <token>`."
      },
      "atServerToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Auto-Tournament-Token",
        "description": "Server-to-API authentication (webhooks, reports, demo uploads, match config fetch). Send `X-Auto-Tournament-Token: <token>`."
      }
    },
...
  },
...
  "paths": {
...
    "/api/demos/{matchSlug}/download/{mapNumber}?": {
      "get": {
        "tags": [
          "Demos"
        ],
        "summary": "GET /api/demos/:matchSlug/download/:mapNumber?",
        "description": "Generated from the router. No hand-written `@openapi` block exists for this endpoint yet, so the request and response shapes are not described — read the handler.",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "matchSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mapNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/demos/{matchSlug}/info": {
      "get": {
        "tags": [
          "Demos"
        ],
        "summary": "GET /api/demos/:matchSlug/info",
        "description": "Generated from the router. No hand-written `@openapi` block exists for this endpoint yet, so the request and response shapes are not described — read the handler.",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "matchSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ]
      }
    },
    "/api/demos/{matchSlug}/status": {
      "get": {
        "tags": [
          "Demos"
        ],
        "summary": "GET /api/demos/:matchSlug/status",
        "description": "Generated from the router. No hand-written `@openapi` block exists for this endpoint yet, so the request and response shapes are not described — read the handler.",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "matchSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ]
      }
    },
    "/api/demos/{matchSlug}/upload": {
      "post": {
        "tags": [
          "Demos"
        ],
        "summary": "POST /api/demos/:matchSlug/upload",
        "description": "Generated from the router. No hand-written `@openapi` block exists for this endpoint yet, so the request and response shapes are not described — read the handler.",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "matchSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "atServerToken": []
          }
        ]
      }
    },
...
  }
}